choro: idGenre: uint → ushort
тип idGenre изменён на ushort

+++ freeLib/src/importthread.cpp
@@ -165,7 +165,7 @@ uint ImportThread::AddBook(qlonglong star, const QString &name, qlonglong id_ser
 
 void ImportThread::AddGenre(uint idBook, const QString &sGenre, uint idLib)
 {
-    qlonglong idGenre = 0;
+    ushort idGenre = 0;
     QString sCorrectGenre = sGenre.toLower();
     sCorrectGenre.replace(' ', '_');
     if(genreKeys_.contains(sCorrectGenre)){
@@ -603,7 +603,7 @@ void ImportThread::process()
     query_.exec(QStringLiteral("SELECT id, keys FROM genre where NOT keys='';"));
     while(query_.next())
     {
-        uint idGenre = query_.value(0).toUInt();
+        ushort idGenre = query_.value(0).toUInt();
         QString sKeys = query_.value(1).toString();
         QStringList kistKeys = sKeys.split(QStringLiteral(";"));
         for(qsizetype i=0; i<kistKeys.size(); i++){
